requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1.2
Pillow==8.0.1
pytz==2020.1
sqlparse==0.4.1
